Script for building PDF manual: run bibtex and makeindex only if needed (as indicated by the existence of .bbl and/or .idx files)
git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@3669 2bfe0521-f11c-4a00-b80e-6202646ff360

+++ doc/manual/mkpdf.sh
@@ -4,7 +4,7 @@ cd `dirname $0`
 
 JOBNAME=manual
 
-MAXRUNS=10
+MAXRUNS=6
 SUCCESS=0
 RUNS=0
 
@@ -15,6 +15,12 @@ rm -f ${JOBNAME}.ind
 until [ ${SUCCESS} = 1 ] ; do
    rm -f ${JOBNAME}.log
    pdflatex ${JOBNAME}
+   grep -q "^No file ${JOBNAME}\.bbl" ${JOBNAME}.log && NEED_BIBTEX=1
+   if [ ${NEED_BIBTEX} = 1 ] ; then
+      bibtex ${JOBNAME}
+      rm -f ${JOBNAME}.log
+      pdflatex ${JOBNAME}
+   fi
    fgrep -q "Rerun to get cross-references right." ${JOBNAME}.log || SUCCESS=1
    RUNS=`expr ${RUNS} + 1`
    if [ ${RUNS} -gt ${MAXRUNS} ] ; then
@@ -28,14 +34,18 @@ until [ ${SUCCESS} = 1 ] ; do
 #   checksum ${JOBNAME}.aux
 done
 
-# finally, run makeindex and latex once
-sed -e 's/^\(\\indexentry{[^@]*\)\\textbar/\1"|/' \
-    -e 's/^\(\\indexentry{[^@]*\)\\underscore/\1_/' \
-    -e 's/^\(\\indexentry{[^@]*\)\\char `\\\\/\1\\\\/' \
-  <manual.idx >manual.idxtmp ; mv manual.idxtmp manual.idx
+if test -f ${JOBNAME}.idx ; then
+  # finally, run makeindex and latex once
+  # replace symbolic representation of | _ \ by the characters themselves,
+  # for correct sorting
+  sed -e 's/^\(\\indexentry{[^@]*\)\\textbar/\1"|/' \
+      -e 's/^\(\\indexentry{[^@]*\)\\underscore/\1_/' \
+      -e 's/^\(\\indexentry{[^@]*\)\\char `\\\\/\1\\\\/' \
+    <manual.idx >manual.idxtmp ; mv manual.idxtmp manual.idx
 
-makeindex ${JOBNAME}
-pdflatex ${JOBNAME}
+  makeindex ${JOBNAME}
+  pdflatex ${JOBNAME}
+fi
 
 exit 0
